test: Split transcoder subs into their own file.
Also add some logging prior to tests.
